requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
blinker==1.4
cachetools==3.1.1
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
cryptography==2.9.2
cycler==0.10.0
cytoolz==0.10.1
dask==2.16.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imagecodecs-lite==2019.12.3
imageio==2.8.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.2.6
jupyterlab-server==1.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
mpi4py==3.0.3
nb-conda-kernels==2.2.3
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pooch==1.1.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.1
scikit-learn==0.22.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tifffile==2020.5.7
toolz==0.10.0
torch==1.4.0
torchvision==0.2.1
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
veusz==3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
